fix: renamed experiment directory
