+++ pysatl_experiment/cli/commands/build_and_run/build_and_run.py
@@ -2,8 +2,8 @@ from click import BadParameter, argument, command
 
 from pysatl_experiment.cli.commands.common.common import read_experiment_data
 from pysatl_experiment.configuration.experiment_data.experiment_data import ExperimentData
-from pysatl_experiment.experiment_new.experiment.experiment import Experiment
-from pysatl_experiment.experiment_new.experiment_steps.experiment_steps import ExperimentSteps
+from pysatl_experiment.experiment.experiment.experiment import Experiment
+from pysatl_experiment.experiment.experiment_steps.experiment_steps import ExperimentSteps
 from pysatl_experiment.factory.critical_value.critical_value import CriticalValueExperimentFactory
 from pysatl_experiment.factory.power.power import PowerExperimentFactory
 from pysatl_experiment.factory.time_complexity.time_complexity import TimeComplexityExperimentFactory
