+++ pysatl_experiment/experiment/step/execution/time_complexity/time_complexity.py
@@ -3,11 +3,11 @@ from dataclasses import dataclass
 
 from line_profiler import profile
 
-from pysatl_experiment.experiment_new.model.experiment_step.experiment_step import IExperimentStep
-from pysatl_experiment.experiment_new.step.execution.common.execution_step_data.execution_step_data import (  # noqa: E501
+from pysatl_experiment.experiment.model.experiment_step.experiment_step import IExperimentStep
+from pysatl_experiment.experiment.step.execution.common.execution_step_data.execution_step_data import (  # noqa: E501
     ExecutionStepData,
 )
-from pysatl_experiment.experiment_new.step.execution.common.hypothesis_generator_data.hypothesis_generator_data import (  # noqa: E501
+from pysatl_experiment.experiment.step.execution.common.hypothesis_generator_data.hypothesis_generator_data import (  # noqa: E501
     HypothesisGeneratorData,
 )
 from pysatl_experiment.parallel.buffered_saver import BufferedSaver
