+++ pysatl_experiment/experiment/step/generation/generation.py
@@ -4,7 +4,7 @@ from line_profiler import profile
 from typing_extensions import override
 
 from pysatl_experiment.experiment.generator import AbstractRVSGenerator
-from pysatl_experiment.experiment_new.model.experiment_step.experiment_step import IExperimentStep
+from pysatl_experiment.experiment.model.experiment_step.experiment_step import IExperimentStep
 from pysatl_experiment.persistence.model.random_values.random_values import IRandomValuesStorage, RandomValuesModel
