+++ pysatl_experiment/experiment/step/report_building/time_complexity/time_complexity.py
@@ -6,7 +6,7 @@ from typing_extensions import override
 
 from pysatl_experiment.configuration.criteria_config.criteria_config import CriterionConfig
 from pysatl_experiment.configuration.model.report_mode.report_mode import ReportMode
-from pysatl_experiment.experiment_new.model.experiment_step.experiment_step import IExperimentStep
+from pysatl_experiment.experiment.model.experiment_step.experiment_step import IExperimentStep
 from pysatl_experiment.persistence.model.time_complexity.time_complexity import (
     ITimeComplexityStorage,
     TimeComplexityQuery,
