+++ pysatl_experiment/factory/critical_value/critical_value.py
@@ -3,15 +3,15 @@ from pysatl_criterion.persistence.model.limit_distribution.limit_distribution im
     LimitDistributionQuery,
 )
 from pysatl_experiment.configuration.experiment_data.critical_value.critical_value import CriticalValueExperimentData
-from pysatl_experiment.experiment_new.step.execution.common.hypothesis_generator_data.hypothesis_generator_data import (  # noqa: E501
+from pysatl_experiment.experiment.step.execution.common.hypothesis_generator_data.hypothesis_generator_data import (  # noqa: E501
     HypothesisGeneratorData,
 )
-from pysatl_experiment.experiment_new.step.execution.critical_value.critical_value import (
+from pysatl_experiment.experiment.step.execution.critical_value.critical_value import (
     CriticalValueExecutionStep,
     CriticalValueStepData,
 )
-from pysatl_experiment.experiment_new.step.generation.generation import GenerationStep, GenerationStepData
-from pysatl_experiment.experiment_new.step.report_building.critical_value.critical_value import (
+from pysatl_experiment.experiment.step.generation.generation import GenerationStep, GenerationStepData
+from pysatl_experiment.experiment.step.report_building.critical_value.critical_value import (
     CriticalValueReportBuildingStep,
 )
 from pysatl_experiment.factory.model.abstract_experiment_factory.abstract_experiment_factory import (
